requirements.txt
addict
future
lmdb
numpy
matplotlib
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7
torchvision
tqdm
yapf
easydict
icecream
einops
timm
